+++ packages/Angular/Generic/gantt/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @memberjunction/ng-gantt
 
+## 6.0.0
+
+### Patch Changes
+
+- @memberjunction/global@6.0.0
+
+## 5.51.0
+
+### Patch Changes
+
+- @memberjunction/global@5.51.0
+
+## 5.50.0
+
+### Patch Changes
+
+- @memberjunction/global@5.50.0
+
 ## 5.49.0
 
 ### Patch Changes